Move new voting tests to a separate file as suggested.
